Split AMTIncrementalWriteSuite into a shared base and topical suites
